Initial Commit
This is the initial commit. Be aware that this current neural network can only perform classification using MNIST dataset of images. When using  the code, download the MNIST data set and put it in the data folder. You can download the data set using this link: https://github.com/cvdfoundation/mnist